Major refactoring

+++ myexaminer/src/main/java/com/myexaminer/security/payload/JwtResponse.java
@@ -18,13 +18,13 @@ public class JwtResponse {
     @Builder.Default
     private String type = "Bearer";
 
-    private Integer id;
+    private Long id;
 
     private String email;
 
     private List<String> roles;
 
-    public JwtResponse(String token, Integer id, String email, List<String> roles) {
+    public JwtResponse(String token, Long id, String email, List<String> roles) {
         this.token = token;
         this.id = id;
         this.email = email;
